+++ package.json
@@ -146,12 +146,13 @@
     "compile": "webpack --mode none",
     "watch": "webpack --mode none --watch",
     "package": "vsce package",
-    "pretest": "tsc -p src/test/tsconfig.test.json",
-    "test": "ts-node ./src/test/runTest.ts",
+    "pretest": "tsc -p test/tsconfig.test.json",
+    "test": "ts-node test/runTest.ts",
     "test:docker": "docker image build . -t evergreen-test"
   },
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^0.1.3",
+    "@types/assert": "^1.4.3",
     "@types/cron": "^1.7.1",
     "@types/execa": "^0.9.0",
     "@types/get-stream": "^3.0.2",
