--- src/test/mocha.opts
@@ -1,4 +1,0 @@
---require ts-node/register # replace with ts-node/register/transpile-only if you have custom types
---require source-map-support/register
---recursive
-**/*.test.ts
