+++ test/mocha.opts
@@ -1,0 +1,6 @@
+--require ts-node/register
+--watch-extensions ts
+--require source-map-support/register
+--recursive
+--reporter spec
+--timeout 5000
