+++ test/mocks/vscode.mock.ts
