+++ test/suite/extension/extension.test.ts
