+++ test/tsconfig.test.json
@@ -1,13 +1,10 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../tsconfig.json",
   "compilerOptions": {
     "declaration": false,
     "strict": false,
     "sourceMap": true,
     "noImplicitAny": false,
-    "types": [
-      "node",
-      "mocha"
-    ],
+    "types": ["node", "mocha", "assert"]
   }
 }
